+++ tests/test_plot/test_int_bom.py
@@ -48,9 +48,10 @@ pytest-3 --log-cli-level debug
 
 """
 
-import os
-import logging
 from base64 import b64decode
+import logging
+import os
+import pytest
 from . import context
 from kibot.misc import EXIT_BAD_CONFIG
 
@@ -1335,6 +1336,24 @@ def test_int_bom_variant_t2s(test_dir):
 def test_int_bom_variant_t2if(test_dir):
     """ IBoM variants test full """
     prj = 'kibom-variant_3'
+    ctx = context.TestContextSCH(test_dir, prj, 'int_bom_var_t2i_csv', BOM_DIR)
+    ctx.run()
+    rows, header, info = ctx.load_csv(prj+'-bom.csv')
+    ref_column = header.index(REF_COLUMN_NAME)
+    check_kibom_test_netlist(rows, ref_column, 1, ['C1', 'C2'], ['R1', 'R2'])
+    rows, header, info = ctx.load_csv(prj+'-bom_[2].csv')
+    check_kibom_test_netlist(rows, ref_column, 1, ['C1', 'C2'], ['R1', 'R2'])
+    rows, header, info = ctx.load_csv(prj+'-bom_(production).csv')
+    check_kibom_test_netlist(rows, ref_column, 2, ['C1'], ['R1', 'R2', 'C2'])
+    rows, header, info = ctx.load_csv(prj+'-bom_(test).csv')
+    check_kibom_test_netlist(rows, ref_column, 2, ['R2'], ['R1', 'C1', 'C2'])
+    ctx.clean_up(keep_project=True)
+
+
+@pytest.mark.skipif(context.ki5(), reason="needs KiCad 6 text variables")
+def test_int_bom_variant_t2it(test_dir):
+    """ IBoM variants test full, here we expand KiCad 6 variables """
+    prj = 'kibom-variant_3_txt'
     ctx = context.TestContextSCH(test_dir, prj, 'int_bom_var_t2i_csv', BOM_DIR)
     ctx.run()
     rows, header, info = ctx.load_csv(prj+'-bom.csv')